requirements.txt
fastapi
uvicorn
python-multipart
# Pinning to known working versions to avoid ImportError: cannot import name 'segformer'
marker-pdf==0.2.14
surya-ocr==0.4.14
grobid-client-python
# Force CPU versions to save space and prevent build timeouts
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
numpy<2.0.0
beautifulsoup4
lxml
opencv-python-headless
transformers==4.44.2